% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_country.R
\name{convert_country}
\alias{convert_country}
\title{Convert country names}
\usage{
convert_country(name, type, ...)
}
\arguments{
\item{name}{Country name to be converted}

\item{type}{Can take the values "full", "ISO2", and "ISO3"}
}
\description{
This is a function allowing for the conversion of country names from one
format to another. It uses the list of 31 EU/EEA countries.
}
