% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting.R
\name{createandview}
\alias{createandview}
\title{Create and view R documentation}
\usage{
createandview(filename, type = "html", ...)
}
\arguments{
\item{type}{Defaults to "html"}

\item{file}{Name of the Rdfile}
}
\description{
This is a wrapper that uses the [devtools::document()] function to create the
documentation, including Rd files, followed by the [Rdpack::viewRd()] function
to show the documentation for the specified file.
It is intended for use when creating R packages.
}
\examples{
createandview("my_function_name.Rd")
}
