% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_wrappers.R
\name{data_and_info}
\alias{data_and_info}
\title{Get both data set and information on it}
\usage{
data_and_info(dataset = "aact_ali01", ...)
}
\arguments{
\item{dataset}{code of Eurostat dataset to retrieve}
}
\description{
This is a wrapper for the Eurostat package which retrieves both the data and
information about that data and returns it as a list
}
\examples{
data_and_info()
}
