% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot}
\alias{dotplot}
\title{This is a function for plotting dotplots the right way round. It primarily uses
[ggplot2::ggplot()] with a bit of [dplyr::inner_join()].}
\usage{
dotplot(data = cars, var = "speed")
}
\arguments{
\item{data}{Dataset to use, defaults to cars}

\item{var}{Desired variable for plotting, defaults to "speed"}
}
\description{
This is a function for plotting dotplots the right way round. It primarily uses
[ggplot2::ggplot()] with a bit of [dplyr::inner_join()].
}
\examples{
dotplot(data = cars, var = "speed")
}
