% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get countries from interinstitutional style guide}
\usage{
get_countries()
}
\description{
Get the list of countries from the Annexes from the Interinstitutional
style guide and save countries for use with convert_country.
}
\examples{
get_countries()
}
\seealso{
[convert_country()]
}
