% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_wrappers.R
\name{least_missings}
\alias{least_missings}
\title{Find dataset from year with least missings}
\usage{
least_missings(dataset = "ilc_lvho05a", ...)
}
\arguments{
\item{dataset}{code of Eurostat dataset to retrieve}
}
\description{
Find dataset from year with least missings
}
\examples{
least_missings()
}
