% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_wrappers.R
\name{list_all_sets}
\alias{list_all_sets}
\title{List all the available datasets}
\usage{
list_all_sets()
}
\arguments{
\item{dataset}{code of Eurostat dataset to retrieve}
}
\description{
This is a wrapper for the Eurostat package which prints a list of the entire
list of available data sets
}
\examples{
list_all_sets()
}
