% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_data}
\alias{save_data}
\title{Save data in various formats}
\usage{
save_data(obj, filename = paste0("data", sprintf(fmt = "\%03d", 1)),
  location = "Data/", types = c("rds", "csv"))
}
\arguments{
\item{obj}{object to be saved}

\item{filename}{name of object}

\item{location}{location object saves to (defaults to subfolder called Data)}

\item{types}{output formats}
}
\description{
This is a wrapper for saving data to local machines.
}
