% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_plot}
\alias{save_plot}
\title{Save plots in various formats}
\usage{
save_plot(fig, filename = "Rplot\%03d", location = "Figures/",
  width = 4.8, height = 4.8, types = c("bmp", "png", "jpeg", "tiff",
  "pdf"))
}
\arguments{
\item{fig}{plot}

\item{filename}{name of plot}

\item{location}{location plot saves to (defaults to subfolder called Figures)}

\item{width}{width of figure defaults to 4.8 (corresponding to 480px
for those exports which have default units px)}

\item{height}{height of figure defaults to 4.8 (corresponding to 480px
for those exports which have default units px)}

\item{types}{output formats}
}
\description{
This is a wrapper for the outputs from grDevices saving plots in all manner of
formats.
}
\examples{
p <- ggplot2::qplot(c(1, 2, 3, 4))
eurosurveillance_types <- c("pdf", "eps", "wmf", "emf", "svg")
save_plot(fig = p, filename = "test", types = eurosurveillance_types)
}
